Add Contact Section and refactor Skills code
