+++ src/Components/Footer/Footer.css
@@ -1,0 +1,15 @@
+footer {
+  position: absolute;
+  left: 0;
+
+  background-color: rgb(26, 26, 70);
+  width: 100%;
+  text-align: center;
+
+  height: 10vh;
+
+  font-size: 2.5vh;
+  font-weight: bold;
+  overflow-x: hidden;
+  z-index: -999;
+}
