+++ src/Components/Footer/Footer.tsx
@@ -1,0 +1,11 @@
+import "./Footer.css";
+
+const Footer = () => {
+  return (
+    <footer>
+      <p>Built and Designed by Henrique Ferreira</p>
+    </footer>
+  );
+};
+
+export default Footer;
