+++ src/Components/ProjectItem/ProjectItem.css
@@ -26,22 +26,26 @@
 
   font-size: 2.8vh;
   font-weight: 300;
+  
+  text-align: justify;
 
   padding: 10px;
 }
 
-.project-description span {
-  display: inline-block;
+.project-tags {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: start;
+  width: 100%;
+  margin-bottom: 20px;
+}
+
+.project-tags span {
   border: 1px solid rgba(127, 255, 212);
   margin: 1vh;
   color: rgb(127, 255, 212);
   padding: 3px 10px;
   border-radius: 1.5vh;
-}
-
-.project-tags {
-  max-width: 500px;
-  margin-bottom: 20px;
 }
 
 .project-description button {
@@ -67,11 +71,14 @@
   border-radius: 1.5vh;
 }
 
-@media only screen and (max-width: 820px) {
+@media only screen and (max-width: 990px) {
   .project {
     flex-direction: column;
 
     text-align: center;
+    justify-content: center;
+  }
+  .project-tags {
     justify-content: center;
   }
   .project-description {
